+++ Tests/MapboxDirectionsTests/IntersectionTests.swift
@@ -19,6 +19,10 @@ class IntersectionTests: XCTestCase {
                     "type": "toll_booth",
                     "name": "test toll booth"
                 ],
+                "railway_crossing": true,
+                "traffic_signal": true,
+                "stop_sign": false,
+                "yield_sign": false
             ],
             [
                 "out": 1,
@@ -59,6 +63,10 @@ class IntersectionTests: XCTestCase {
             XCTAssertEqual(intersection.headings, [80.0])
             XCTAssertEqual(intersection.location, LocationCoordinate2D(latitude: 52.508068, longitude: 13.426579))
             XCTAssertEqual(intersection.outletMapboxStreetsRoadClass, MapboxStreetsRoadClass.streetLimited)
+            XCTAssertEqual(intersection.railroadCrossing, true)
+            XCTAssertEqual(intersection.trafficSignal, true)
+            XCTAssertEqual(intersection.stopSign, false)
+            XCTAssertEqual(intersection.yieldSign, false)
         }
         
         intersections = [
@@ -76,7 +84,11 @@ class IntersectionTests: XCTestCase {
                          tunnelName: nil,
                          restStop: nil,
                          isUrban: nil,
-                         outletMapboxStreetsRoadClass: .streetLimited),
+                         outletMapboxStreetsRoadClass: .streetLimited,
+                         railroadCrossing: true,
+                         trafficSignal: true,
+                         stopSign: false,
+                         yieldSign: false),
             Intersection(location: LocationCoordinate2D(latitude: 52.508022, longitude: 13.426688),
                          headings: [30.0, 120.0, 300.0],
                          approachIndex: 2,